Release: 0.0.1-dev011
Today I finished moving all the logic to perform routing to FSMRouter class, which makes it possible to reuse that piece of code later to support goto outside fsm state.

This, however, is not ready yet. Support of routing outside fsm is crucial for cases like routing on cron timer, or routing from private messages state to callback query state.